Fixed response classes, some issues with survey.
