+++ api/models/survey.py
@@ -64,6 +64,11 @@ class Survey(Base):
             self.isPublic = False
         if isSurveySentAutomatically is not None:
             self.isSurveySentAutomatically = isSurveySentAutomatically
+        for allowed_respondent in self.respondents:
+            allowed_respondent.delete_respondent()
+        for response in self.responses:
+            response.delete_response()
+
 
     def generate_hash(self):
         str = string.ascii_lowercase + string.ascii_uppercase
@@ -73,7 +78,7 @@ class Survey(Base):
         return self.surveyHash == surveyHash
 
     def is_published(self):
-        return datetime.datetime.now(datetime.timezone.utc).timestamp() <= self.startDate.timestamp() and datetime.datetime.now(datetime.timezone.utc).timestamp() <= self.endDate.timestamp()
+        return datetime.datetime.now(datetime.timezone.utc).timestamp() >= self.startDate.timestamp() and datetime.datetime.now(datetime.timezone.utc).timestamp() <= self.endDate.timestamp()
 
     def serialize(self):
         return {
@@ -174,3 +179,152 @@ class Survey(Base):
         form_builder['models'] = models
         survey_json['data']['formBuilder'] = form_builder
         return survey_json
+
+
+class Respondents(Base):
+    __tablename__ = 'respondents'
+    id = Column(Integer, primary_key=True, index=True)
+    email = Column(String(255), nullable=False)
+    surveys = relationship("AllowedRespondents", back_populates="respondent", cascade="all, delete-orphan")
+    responses = relationship("Responses", back_populates="respondent", cascade="all, delete-orphan")
+
+    def __init__(self, email):
+        self.email = email
+
+    def serialize(self):
+        return {
+            'id': self.id,
+            'email': self.email
+        }
+
+    def add_respondent(self):
+        if not session.query(Respondents).filter_by(email=self.email).first():
+            session.add(self)
+            session.commit()
+
+    def delete_respondent(self):
+        session.delete(self)
+        session.commit()
+
+    @staticmethod
+    def get_respondent(email):
+        return session.query(Respondents).filter_by(email=email).first()
+
+
+class Responses(Base):
+    __tablename__ = 'responses'
+    id = Column(Integer, primary_key=True, index=True)
+    respondentId = Column(Integer, ForeignKey(Respondents.id), nullable=False)
+    surveyId = Column(Integer, ForeignKey(Survey.id), nullable=False)
+    survey = relationship("Survey", back_populates="responses")
+    respondent = relationship("Respondents", back_populates="responses")
+    answers = relationship("Answers", back_populates="response", cascade="all, delete-orphan")
+
+    def __init__(self, respondentId, surveyId):
+        self.respondentId = respondentId
+        self.surveyId = surveyId
+
+    def add_response(self):
+        session.add(self)
+        session.commit()
+
+    def delete_response(self):
+        session.delete(self)
+        session.commit()
+
+    @staticmethod
+    def get_responses(surveyId):
+        return session.query(AllowedRespondents).filter_by(surveyId=surveyId)
+
+
+from models.question import Question, ScaleQuestion, OpenAnswerQuestion, MultipleChoiceQuestion, AnswerOption
+
+
+class Answers(Base):
+    __tablename__ = 'answers'
+    id = Column(Integer, primary_key=True, index=True)
+    responseId = Column(Integer, ForeignKey(Responses.id), nullable=False)
+    response = relationship("Responses", back_populates="answers")
+    open_answers = relationship("OpenAnswers", back_populates="answer", cascade="all, delete-orphan")
+    scale_answers = relationship("ScaleAnswers", back_populates="answer", cascade="all, delete-orphan")
+    choice_answers = relationship("ChoiceAnswers", back_populates="answer", cascade="all, delete-orphan")
+
+    def __init__(self, responseId):
+        self.responseId = responseId
+
+    def add_answer(self):
+        session.add(self)
+        session.commit()
+
+
+class OpenAnswers(Base):
+    __tablename__ = 'open_answers'
+    id = Column(Integer, primary_key=True, index=True)
+    answerId = Column(Integer, ForeignKey(Answers.id), nullable=False)
+    answer = relationship("Answers", back_populates="open_answers")
+    open_answer_question_id = Column(Integer, ForeignKey(OpenAnswerQuestion.id), nullable=False)
+    open_question = relationship("OpenAnswerQuestion", back_populates="open_answers")
+    answer_text = Column(String(255), nullable=False)
+
+    def __init__(self, answerId, open_answer_question_id, answer_text):
+        self.answerId = answerId
+        self.open_answer_question_id = open_answer_question_id
+        self.answer_text = answer_text
+
+    def add_answer(self):
+        session.add(self)
+        session.commit()
+
+
+class ScaleAnswers(Base):
+    __tablename__ = 'scale_answers'
+    id = Column(Integer, primary_key=True, index=True)
+    answerId = Column(Integer, ForeignKey(Answers.id), nullable=False)
+    answer = relationship("Answers", back_populates="scale_answers")
+    scale_answer_question_id = Column(Integer, ForeignKey(ScaleQuestion.id), nullable=False)
+    scale_question = relationship("ScaleQuestion", back_populates="scale_answers")
+    value = Column(Integer, nullable=False)
+
+    def __init__(self, answerId, scale_answer_question_id, value):
+        self.answerId = answerId
+        self.scale_answer_question_id = scale_answer_question_id
+        self.value = value
+
+    def add_answer(self):
+        session.add(self)
+        session.commit()
+
+
+class ChoiceAnswers(Base):
+    __tablename__ = 'choice_answers'
+    id = Column(Integer, primary_key=True, index=True)
+    answerId = Column(Integer, ForeignKey(Answers.id), nullable=False)
+    answer = relationship("Answers", back_populates="choice_answers")
+    multiple_choice_questionsId = Column(Integer, ForeignKey(MultipleChoiceQuestion.id), nullable=False)
+    choice_question = relationship("MultipleChoiceQuestion", back_populates="choice_answers")
+    answer_options_choice = relationship("AnswerOptionsChoiceAnswer", back_populates="choiceAnswer", cascade="all, delete-orphan")
+
+    def __init__(self, answerId, multiple_choice_questionsId):
+        self.answerId = answerId
+        self.multiple_choice_questionsId = multiple_choice_questionsId
+
+    def add_answer(self):
+        session.add(self)
+        session.commit()
+
+
+class AnswerOptionsChoiceAnswer(Base):
+    __tablename__ = 'answer_options_choice_answer'
+    id = Column(Integer, primary_key=True, index=True)
+    choiceAnswersId = Column(Integer, ForeignKey(ChoiceAnswers.id), nullable=False)
+    choiceAnswer = relationship("ChoiceAnswers", back_populates="answer_options_choice")
+    answer_optionId = Column(Integer, ForeignKey(AnswerOption.id), nullable=False)
+    answer_option = relationship("AnswerOption", back_populates="answer_options_choice_answers")
+
+    def __init__(self, choiceAnswersId, answer_optionId):
+        self.choiceAnswersId = choiceAnswersId
+        self.answer_optionId = answer_optionId
+
+    def add_answer(self):
+        session.add(self)
+        session.commit()
