+++ api/resources/survey.py
@@ -139,14 +139,17 @@ class GetSurvey(Resource):
                 if not is_allowed(User.find_by_id(current_user_id)):
                     return {'message': 'You are not allowed to access the website.'}, 403
                 if survey.surveyOwner == current_user_id:
-                    allow = True
+                    return jsonify(survey.get_json())
+        except Exception:
+            allow = False
+        if survey.is_published():
             if surveyHash is not None:
                 if surveyHash == survey.surveyHash:
                     allow = True
             if survey.isPublic:
                 allow = True
-        except Exception:
-            allow = False
         if not allow:
             return {'message': 'You are not allowed to access this survey.'}, 403
-        return jsonify(survey.get_json())
+        survey_json = survey.get_json()
+        survey_json['config'] = dict()
+        return jsonify(survey_json)
